requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
twilio==8.10.0
openai==1.3.5
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
